[refactor] adding comparison test file
