+++ test/distanceWithTime.test.ts
@@ -2,8 +2,6 @@ import {distanceWithTime} from "../src/distanceWithTime";
 import {checkpointsTimestamps, getRailwayReference, getTimedRun, path1, path2} from "./test.features";
 import {expect} from "chai";
 import {clone} from "@turf/turf";
-import { pathDistance } from "@indoor-analytics/path-distance";
-import { ErrorVector } from "@indoor-analytics/entities";
 
 describe('distanceWithTime', () => {
     it ('should throw with empty timestamps array', () => {
@@ -21,22 +19,5 @@ describe('distanceWithTime', () => {
         faultyPath.properties!.locationsTimestamps = [0, 1, 2, 3];
         const getDistance = () => distanceWithTime(getRailwayReference(), faultyPath, checkpointsTimestamps);
         expect(getDistance).to.throw(RangeError, 'Compared path timestamps length must match its locations count.');
-    });
-
-    it ('should be more accurate than pathDistance method with run having many locations at the beginning of reference path', () => {
-        const referencePath = getRailwayReference();
-        const comparedPath = getTimedRun();
-
-        const classicVectors = pathDistance(referencePath, comparedPath);
-        let total = 0;
-        classicVectors.forEach(v => total += v.distance);
-        const classicVectorsAverageError = total/classicVectors.length;
-
-        total = 0;
-        const timeVectors = distanceWithTime(referencePath, comparedPath, checkpointsTimestamps);
-        timeVectors.forEach(v => total += v.distance);
-        const timeVectorsAverageError = total/timeVectors.length;
-
-        expect(timeVectorsAverageError).to.be.greaterThan(classicVectorsAverageError);
     });
 });
